function for adding random paths and delete function for random paths implemented

+++ app.js
@@ -19,18 +19,27 @@ const taskschema = new mongoose.Schema({
     }   
 })
 
-const Task = mongoose.model('task',taskschema)
 
+const pathschema = new mongoose.Schema({
+    name : String,
+    item : [taskschema]
+})
+
+const Task = mongoose.model('task',taskschema)
+const Path = mongoose.model('path',pathschema)
+var date2 = getAmericanDate()
 
 app.get('/',async(req,res)=>{
     const todo = await Task.find()
-    var date2 = getAmericanDate()
-    console.log(date2)
     res.render('index',{todo:todo,date:date2});
 })
 
+
+
 app.post('/',async(req,res)=>{
     const task1 = new Task({task:req.body.newItem})
+    if(date2 === req.body.list){
+        
     try{
         await task1.save();
         res.redirect('/')
@@ -39,19 +48,55 @@ app.post('/',async(req,res)=>{
             res.redirect('/') 
         }
     }
+    }else{
+        const list = await Path.findOne({name:req.body.list})
+        console.log(list)
+        list.item.push(task1)
+        await list.save()
+        res.redirect('/'+req.body.list) 
+    }
+    
     
     
 })
+
+
 app.post('/delete',async (req,res)=>{
     const dltitem = req.body.checkbox;
-    await Task.findByIdAndRemove(dltitem)
-    res.redirect('/')
+    const dltpath = req.body.hidden;
+    if(date2 === dltpath){
+        await Task.findByIdAndRemove(dltitem)
+        res.redirect('/')
+    }else{
+        const dltpathitem = await Path.findOneAndUpdate({name:dltpath},{$pull:{item:{_id:dltitem}}})
+        res.redirect('/'+dltpath)
+    }
+    
 })
 
-// app.get('/:pathid',(req,res)=>{
-//     const pathid = req.params.pathid
 
-// })
+
+const defualitems=['welcome to todolist']
+app.get('/:pathid',async(req,res)=>{
+    const pathid = req.params.pathid
+    try{
+        const hello=await Path.findOne({name:pathid})
+        if(hello ===null){
+            const path1 = new Path({name:pathid})
+            await path1.save()
+            res.redirect('/'+Path.name) 
+        }else{
+            const pathfinder = hello.item
+            res.render('index',{todo:pathfinder,date:pathid});
+        }
+        
+        
+    }catch(err){
+        if(err){
+            
+        }
+    }
+})
 
 app.listen(3000,()=>{
     console.log('app running succesfully')
